resolved refactoring bugs in query, compute, dashboard
